Post Editor now reliably shows the copied content
The idea here is that a copy is new but has content.
As _isNew isn't changed immediately after setting the editor content,
setEditorContent fires only the first time and would fire again only if
the post isNew and the content is empty (both conditions are not likely
to happen again after the copy).